Fix timezone issue with next_future_time calculation.
By default, unless specified, the timezone is set to UTC for the first run_at date. However, when the next run_at was calculated, the run_at would look at the server time, rather using the same timezone that the calculated run_at was in. This led to some errors if the server was not in UTC time.